backend/requirements.txt
fastapi==0.115.8
uvicorn[standard]==0.34.0
sqlalchemy==2.0.38
psycopg2-binary==2.9.10
alembic==1.14.1
pydantic==2.10.6
pydantic-settings==2.8.1
python-jose[cryptography]==3.3.0
httpx==0.28.1
